requirements/test.txt
-r base.txt

factory-boy==2.11.1
coveralls==1.1
flake8==3.6.0
coverage==4.4.1
mock==2.0.0
django-casper==0.0.4
pytest==3.1.0
pytest-django==3.1.2
